fix: 학년, 요일 컴포넌트 분리해서 렌더링 최적화
